Create Servlet Utility to handle common Servlet operations (#55)
* Refactor with abstract class

* Move static variable

* Create a servlet utility file

* Refactor tests

* Move constant from authentication utility to servlet utility

* Documentation changes

* Documentation changes

* Refactor authentication utility to use method for verification

* Lint

* Helper function for getting tasks

* overload get and post methods to handle credential verification

* Lint

* Clean-up

* Remove useless comment

* Use static variables where appropriate

* Remove redundant comment

* Update incorrect comment

* Add assertion

* Create default implementation for post and get

* PR comments

* Fix/Justify scope of doGet and doPost where needed

* Correct scope to allow for unit testing of doGet/doPost

* Undo helper function

* Add assertion messages

* Fix logging of cookies

* Fix static var in test
