+++ src/main/java/com/google/sps/servlets/TasksServlet.java
@@ -28,10 +28,7 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-/**
- * Serves selected information from the User's Tasks Account. TODO: Create Servlet Utility to handle
- * common functions (Issue #26) TODO: Implement Post (Issue #53)
- */
+/** Serves selected information from the User's Tasks Account. TODO: Implement Post (Issue #53) */
 @WebServlet("/tasks")
 public class TasksServlet extends AuthenticatedHttpServlet {
   /**
